Move ViewModifier  (#108)
Reorganizes ViewModifier (previously ViewExt) under the modifier module
