+++ src/view.rs
@@ -19,7 +19,6 @@ pub mod scroll_view;
 pub mod shape;
 mod spacer;
 mod text;
-mod view_ext;
 mod view_that_fits;
 mod vstack;
 mod zstack;
@@ -41,7 +40,6 @@ pub use scroll_view::ScrollView;
 pub use spacer::Spacer;
 pub(crate) use text::WhitespaceWrap;
 pub use text::{HorizontalTextAlignment, Text};
-pub use view_ext::ViewExt;
 pub use view_that_fits::{FitAxis, ViewThatFits};
 pub use vstack::VStack;
 pub use zstack::ZStack;
@@ -50,12 +48,13 @@ pub use zstack::ZStack;
 pub mod prelude {
     pub use super::aspect_ratio::{ContentMode, Ratio};
     pub use super::button::Seal;
+    pub use super::modifier::ViewModifier;
     pub use super::{padding::Edges, FitAxis, HorizontalTextAlignment};
     #[cfg(feature = "embedded-graphics")]
     pub use super::{AsDrawable, Image};
     pub use super::{
         Button, Divider, EmptyView, ForEach, GeometryReader, HStack, Lens, ScrollView, Spacer,
-        Text, VStack, View, ViewExt, ViewLayout, ViewThatFits, ZStack,
+        Text, VStack, View, ViewLayout, ViewThatFits, ZStack,
     };
     pub use crate::animation::Animation;
     pub use crate::layout::{Alignment, HorizontalAlignment, VerticalAlignment};
